rename: folder orodomop_widgets -> orodomop
