+++ lib/widgets/orodomop/timer_control_row.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:orodomop/providers/timer_provider.dart';
 import 'package:orodomop/models/timer_state.dart';
-import 'package:orodomop/widgets/orodomop_widgets/break_button.dart';
+import 'package:orodomop/widgets/orodomop/break_button.dart';
 import 'package:orodomop/widgets/buttons/end_break_button.dart';
 import 'package:orodomop/widgets/buttons/pause_or_resume_button.dart';
 import 'package:orodomop/widgets/buttons/reset_button.dart';
